requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
pandas
openpyxl
pyodbc
python-multipart
python-dotenv
streamlit
psycopg2-binary
